+++ __tests__/services/grpc/world_engine_client_grpc.service.spec.ts
@@ -87,7 +87,6 @@ describe('load scene', () => {
       name: SCENE,
       capabilities,
       sessionToken,
-      user,
     });
 
     const loadedAgents = result.getAgentsList();